Code deduplication and cleanup
Moved loop handling, exception handling and program
running methods of the gui tools to the utils file.

fixes QubesOS/qubes-issues#5342
